Nest classes and source files within Raindar namespace
