+++ spec/raindar/url_cache_spec.rb
@@ -1,9 +1,9 @@
-require 'url_cache'
+require 'raindar/url_cache'
 
-describe UrlCache do
+describe Raindar::UrlCache do
   describe '.cached_radar_url' do
     let(:radar_code) { "radar_code" }
-    let(:url_cache) { UrlCache.new instance_double(Redis, get: 'url') }
+    let(:url_cache) { Raindar::UrlCache.new instance_double(Redis, get: 'url') }
 
     it 'caches the url returned by the radar generator' do
       expect(url_cache.cached_radar_url(radar_code)).to eq 'url'
